refactor: transform all css files in scss files
