+++ src/components/education/education.module.scss
