+++ src/components/education/education.module.scss.d.ts
