+++ src/components/footer/footer.module.scss.d.ts
