+++ src/components/footer/footer.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "./footer.module.css";
+import styles from "./footer.module.scss";
 
 const Footer = () => (
   <footer className={styles.footer}>
