+++ src/components/projects/work-experience/work-experience.module.scss.d.ts
